Add a comment, move test to trybuild
